+++ DTOs/Payments/clsPaymentDTO.cs
@@ -6,7 +6,7 @@ namespace DTOs.Payments
     {
         public int PaymentID { get; set; }
         public int InvoiceID { get; set; }
-        public float PaymentAmount { get; set; }
+        public decimal PaymentAmount { get; set; }
         public byte PaymentMethodID { get; set; }
         public DateTime PaymentDate { get; set; }
         public int CreatedByUserID { get; set; }
